+++ src/redisearch_rs/timeout/Cargo.toml
@@ -11,5 +11,3 @@ bench = false
 
 [lints]
 workspace = true
-
-[dependencies]
